+++ src/main/java/de/samply/security/ApiKeySecurityConfiguration.java
@@ -2,8 +2,11 @@ package de.samply.security;
 
 
 import de.samply.teiler.TeilerConst;
+import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.annotation.Order;
@@ -37,10 +40,8 @@ public class ApiKeySecurityConfiguration {
    */
   @Bean
   public SecurityFilterChain filterChain(HttpSecurity httpSecurity) throws Exception {
-
     httpSecurity
-        //.cors(Customizer.withDefaults())
-        .cors(cors -> cors.disable())
+        .cors(Customizer.withDefaults())
         .securityMatcher(TeilerConst.REST_PATHS_WITH_API_KEY)
         .csrf().disable()
         .sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS)
@@ -70,13 +71,26 @@ public class ApiKeySecurityConfiguration {
   }
 
   @Bean
-  CorsConfigurationSource corsConfigurationSource() {
+  CorsConfigurationSource corsConfigurationSource(
+      @Value(TeilerConst.CROSS_ORIGINS_SV) String[] crossOrigins) {
     CorsConfiguration configuration = new CorsConfiguration();
-    configuration.setAllowedOrigins(Arrays.asList("http://localhost:9000"));
-    configuration.setAllowedMethods(Arrays.asList("GET","POST", "OPTIONS"));
+    //configuration.setAllowedOrigins(fetchCrossOrigins(crossOrigins));
+    configuration.setAllowedOrigins(Arrays.asList(crossOrigins));
+    configuration.setAllowedMethods(Arrays.asList("GET", "POST", "PUT"));
+    configuration.setAllowedHeaders(
+        Arrays.asList("Authorization", "Cache-Control", "Content-Type", "Origin",
+            TeilerConst.API_KEY_HEADER));
     UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
     source.registerCorsConfiguration("/**", configuration);
     return source;
   }
+
+  /*
+  List<String> fetchCrossOrigins(String crossOrigins) {
+    return (crossOrigins != null) ? Arrays.asList(
+        crossOrigins.split(TeilerConst.CROSS_ORIGINS_SEPARATOR)) : new ArrayList<>();
+  }
+
+   */
 
 }