+++ CHANGELOG.md
@@ -81,6 +81,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Csv Explorer
 - Json Explorer
 - XML Exporter
+- Request Page Size
 
 ### Changed
 - Rename Conversion Template as Converter Template