+++ CHANGELOG.md
@@ -59,3 +59,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Rename Conversion Template as Converter Template
 - First anonym: 1
+- Cors config